Improved SSE Handling and notification handling.
1. Don't send notifications back to the client.
2. Create consistent notifications that can be on a timer or dismissed.

Notifications are not sent back to the client that generated them.

This works even if the same user is logged into more than one
window. Each window generates a 20-character clientId nonce
and sends it with every authenticated request in the X-Client-Id
header. It also supplies the same client ID when register for
notifications. If a client generates a notification, then it is
broadcast to all client IDs except the one that it came from.

This simplifies some controller logic. For example, if you add
a site, the controller adding the note can fetch
the sites collection again immediately. But adding a site also
triggers a site "created" notification. So the original controller
needs to avoid fetching the notes collection a second time. In
this commit, the original control doesn't receive a notification,
so no additional logic is needed.

Furthermore icons are added to the notifications generated by the user to make it clearer whether the notification was created by this or another user.