HGPROF-21 Searching for a user hangs
Remove tornado and implement parallelism using python RQ instead. This
turns out to be tricky since the RQ job ID was used as a foreign key in
a lot of places, under the assumption that "1 search=1 job". With multiple
jobs being used for a single search, its necessary to come up with a
different foreign key, one that I'm calling "tracker ID". In the future,
we should add a model that abstracts the idea of a search and contains
pointers to search results and search archives, but this quick n' dirty
patch gets the job done for now.