HGPROF-1 Finish install cleanup
Fix a bug that I think I introduced during this refactoring: there is supposed
to be a fixture in the database for the generic "an error occurred" image and this
used to be created on-demand if it didn't exist. The problem is if you have 50
workers, then several early workers might each create this error image and then
we have duplicates, which causes future workers to fail when they try to select
just one error image. The fix is to move the fixture creation into the database
script, which requires some fiddling with ownership since the database script
is usually run by root and we need data files to be owned by the application user
(usually "hgprofiler").